Move AST editing into ast/edit.rs
